Started reimplementing the PFA using a bitmap
