--- charlotte_core/src/arch/x86_64/cpu/cpuid/mod.rs
@@ -1,3 +1,0 @@
-extern "C" {
-    pub fn asm_get_vendor_string(dest: &mut [u8; 12]);
-}
