+++ charlotte_core/src/memory/vmm.rs
